added user existence check to contacts service

+++ src/main/java/com/nazipov/merapet/service/impl/ContactsServiceImpl.java
@@ -1,5 +1,9 @@
 package com.nazipov.merapet.service.impl;
 
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
+
 import com.nazipov.merapet.entities.Contact;
 import com.nazipov.merapet.entities.MyUser;
 import com.nazipov.merapet.service.ContactsService;
@@ -11,10 +15,7 @@ import org.springframework.stereotype.Service;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import reactor.core.publisher.MonoOperator;
 
 @Service
 public class ContactsServiceImpl implements ContactsService {
@@ -28,12 +29,18 @@ public class ContactsServiceImpl implements ContactsService {
 
     @Override
     public Mono<Collection<Contact>> retrieveContacts(String userId) {
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
+        }
         return Mono.just(storage.retrieveContacts(userId))
             .doOnNext(signal -> logger.info(String.format("Retrieved all user id: %s contacts", userId)));
     }
 
     @Override
     public Mono<Contact> addContact(String userId, Contact contact) {
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
+        }
         return Mono.just(storage.saveContact(userId, contact))
             .doOnNext(c -> logger.info(String.format("Added contact: %s with name: '%s' for user: %s", 
                     c.getUserId(), 
@@ -45,6 +52,9 @@ public class ContactsServiceImpl implements ContactsService {
 
     @Override
     public Mono<List<Contact>> addContacts(String userId, List<Contact> contacts) {
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
+        }
         return Flux.fromIterable(contacts)
                 .flatMap(c -> addContact(userId, c))
                 .collectList()
@@ -53,6 +63,9 @@ public class ContactsServiceImpl implements ContactsService {
 
     @Override
     public Mono<Contact> editContact(String userId, Contact contact) {
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
+        }
         String contactId = contact.getContactId();
         if (!storage.hasContact(userId, contactId)) {
             return Mono.error(new IllegalArgumentException(String.format("User: %s%n don't have contact: %s", userId, contactId)));
@@ -68,6 +81,9 @@ public class ContactsServiceImpl implements ContactsService {
 
     @Override
     public Mono<List<Contact>> editContacts(String userId, List<Contact> contacts) {
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
+        }
         return Flux.fromIterable(contacts)
                 .flatMap(c -> editContact(userId, c))
                 .collectList()
@@ -76,9 +92,8 @@ public class ContactsServiceImpl implements ContactsService {
 
     @Override
     public Mono<Contact> deleteContact(String userId, Contact contact) {
-        Optional<MyUser> user = storage.retrieveUser(userId);
-        if (user.isEmpty()) {
-            return Mono.error(new IllegalArgumentException(String.format("User: %s doesn't exist", userId)));
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
         }
         
         String contactId = contact.getContactId();
@@ -97,6 +112,9 @@ public class ContactsServiceImpl implements ContactsService {
 
     @Override
     public Mono<List<Contact>> deleteContacts(String userId, List<Contact> contacts) {
+        if (!storage.isUserExistById(userId)) {
+            return Mono.error(new IllegalArgumentException(String.format("There is no user: '%s'", userId)));
+        }
         return Flux.fromIterable(contacts)
                 .flatMap(c -> deleteContact(userId, c))
                 .collectList()
